www/react: Move hasActiveMaster() to buildbot-ui package
